added appveyorapi class
